requirements.txt
# AI Video Creator App - Dependencies
# Core Framework
gradio>=4.44.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Audio Synthesis
edge-tts>=6.1.9
pydub>=0.25.1

# Video Processing
ffmpeg-python>=0.2.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# API & WebSocket
websockets>=12.0
websocket-client>=1.7.0
httpx>=0.26.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# Database & State
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Configuration & Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pillow>=10.2.0

# Logging & Monitoring
loguru>=0.7.2
rich>=13.7.0

# Background Tasks
celery[redis]>=5.3.0
redis>=5.0.1

# Type Hints
typing-extensions>=4.9.0

